Fix initial-step skip in hold_dim_button after turn-on
The "skip next 2 steps" delay used unsigned millis() subtraction,
which underflowed and fired the first dim step immediately — so
holding from OFF jumped past min_brightness instead of starting there.
Use signed comparison so a future last_step_ms_ is honored.

+++ components/hold_dim_button/hold_dim_button.cpp
@@ -56,7 +56,7 @@ void HoldDimButton::loop() {
 
   // while in "hold" mode: execute steps with step_interval_ms_
   if (local_hold) {
-    if (last_step_ms_ == 0 || (now - last_step_ms_) >= step_interval_ms_) {
+    if (last_step_ms_ == 0 || (int32_t)(now - last_step_ms_) >= (int32_t)step_interval_ms_) {
       last_step_ms_ = now;
 
       if (!light_ || !light_->current_values.is_on())
